added features
added search func
added assign hours for assigning hours
added time logged with refresh

fixed backend problem, imported database and added queries in text file xampp

+++ database/migrations/2024_10_21_132933_drop_name_from_users_table.php
@@ -9,12 +9,14 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('name'); // This will remove the 'name' column
-        });
-    }
+    public function up(): void{
+    Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'name')) {
+            $table->dropColumn('name');
+        }
+    });
+}
+
 
     /**
      * Reverse the migrations.
